requirements.txt
asgiref==3.4.1
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.0
jsonfield==3.1.0
Markdown==3.3.6
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.6.0
